Refactor match lowering into partial class
