+++ src/Raven.CodeAnalysis/BoundTree/Lowering/Lowerer.cs
@@ -6,7 +6,7 @@ using Raven.CodeAnalysis.Symbols;
 
 namespace Raven.CodeAnalysis;
 
-internal sealed class Lowerer : BoundTreeRewriter
+internal sealed partial class Lowerer : BoundTreeRewriter
 {
     private readonly ISymbol _containingSymbol;
     private readonly Stack<(ILabelSymbol BreakLabel, ILabelSymbol ContinueLabel)> _loopStack = new();
@@ -64,63 +64,6 @@ internal sealed class Lowerer : BoundTreeRewriter
         }
 
         return new BoundInvocationExpression(node.Method, arguments, receiver, extensionReceiver);
-    }
-
-    public override BoundNode? VisitMatchExpression(BoundMatchExpression node)
-    {
-        var compilation = GetCompilation();
-        var booleanType = compilation.GetSpecialType(SpecialType.System_Boolean);
-        var unitType = compilation.GetSpecialType(SpecialType.System_Unit);
-
-        var scrutinee = (BoundExpression)VisitExpression(node.Expression)!;
-        var scrutineeType = scrutinee.Type ?? compilation.ErrorTypeSymbol;
-        var scrutineeLocal = CreateTempLocal("match_scrutinee", scrutineeType, isMutable: false);
-
-        var statements = new List<BoundStatement>
-        {
-            new BoundLocalDeclarationStatement([
-                new BoundVariableDeclarator(scrutineeLocal, scrutinee)
-            ])
-        };
-
-        var resultType = node.Type ?? compilation.ErrorTypeSymbol;
-        var resultLocal = CreateTempLocal("match_result", resultType, isMutable: true);
-        statements.Add(new BoundLocalDeclarationStatement([
-            new BoundVariableDeclarator(resultLocal, initializer: null)
-        ]));
-
-        var endLabel = CreateLabel("match_end");
-
-        foreach (var arm in node.Arms)
-        {
-            var pattern = (BoundPattern)VisitPattern(arm.Pattern);
-            var guard = (BoundExpression?)VisitExpression(arm.Guard);
-            var expression = (BoundExpression)VisitExpression(arm.Expression)!;
-
-            expression = ApplyConversionIfNeeded(expression, resultType, compilation);
-
-            BoundStatement armResult = new BoundBlockStatement([
-                new BoundExpressionStatement(new BoundLocalAssignmentExpression(resultLocal, expression)),
-                new BoundGotoStatement(endLabel)
-            ]);
-
-            if (guard is not null)
-            {
-                armResult = new BoundIfStatement(guard, armResult, null);
-            }
-
-            var condition = new BoundIsPatternExpression(
-                new BoundLocalAccess(scrutineeLocal),
-                pattern,
-                booleanType);
-
-            statements.Add(new BoundIfStatement(condition, armResult, null));
-        }
-
-        statements.Add(CreateLabelStatement(endLabel));
-        statements.Add(new BoundExpressionStatement(new BoundLocalAccess(resultLocal)));
-
-        return new BoundBlockExpression(statements, unitType);
     }
 
     public override BoundNode? VisitIfStatement(BoundIfStatement node)
